+++ packages/wallet/frontend/src/components/dialogs/UserCardPINDialog.tsx
@@ -2,7 +2,7 @@ import { Dialog, Transition } from '@headlessui/react'
 import { Fragment, useState } from 'react'
 import type { DialogProps } from '@/lib/types/dialog'
 import { cardServiceMock, changePinSchema, IUserCard } from '@/lib/api/card'
-import { UserCardFront } from './UserCard'
+import { UserCardFront } from '@/components/userCards/UserCard'
 import { Button } from '@/ui/Button'
 import { useZodForm } from '@/lib/hooks/useZodForm'
 import { Form } from '@/ui/forms/Form'
